Ajout du bouton publier sur la page de détail d'une fiche détection.
