+++ sv/tests/test_fichedetection_visibilite_update.py
@@ -9,17 +9,8 @@ from core.models import Structure, Visibilite
 from core.constants import BSV_STRUCTURE, MUS_STRUCTURE, AC_STRUCTURE
 
 """
-Les tests vérifient :
-- l'accès à une fiche de détection en fonction de sa visibilité et de l'agent connecté
-- l'accès à la liste de fiches détection en fonction de la visibilité des fiches et de l'agent connecté
-- la modification de la visibilité d'une fiche de détection en de sa visibilité et de l'agent connecté
+Les tests vérifient la modification de la visibilité d'une fiche de détection en de sa visibilité et de l'agent connecté
 """
-
-
-def _update_visibilite_fiche_detection(fiche_detection: FicheDetection, visibilite_libelle: str):
-    """Met à jour la visibilité de la fiche de détection donnée."""
-    fiche_detection.visibilite = visibilite_libelle
-    fiche_detection.save()
 
 
 def test_fiche_detection_visibilite_brouillon(live_server, page: Page, form_elements: FicheDetectionFormDomElements):
@@ -27,6 +18,9 @@ def test_fiche_detection_visibilite_brouillon(live_server, page: Page, form_elem
     page.goto(f"{live_server.url}{reverse('fiche-detection-creation')}")
     page.get_by_role("button", name="Enregistrer le brouillon").click()
     expect(page.get_by_text("brouillon")).to_be_visible()
+    page.wait_for_timeout(600)
+    fiche_detection = FicheDetection.objects.get()
+    assert fiche_detection.visibilite == Visibilite.BROUILLON
 
 
 def test_fiche_detection_visibilite_local(live_server, page: Page, form_elements: FicheDetectionFormDomElements):
@@ -34,90 +28,9 @@ def test_fiche_detection_visibilite_local(live_server, page: Page, form_elements
     page.goto(f"{live_server.url}{reverse('fiche-detection-creation')}")
     page.get_by_role("button", name="Publier").click()
     expect(page.get_by_text("local")).to_be_visible()
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize("visibilite_libelle", [Visibilite.BROUILLON, Visibilite.LOCAL, Visibilite.NATIONAL])
-def test_agent_in_structure_createur_can_view_fiche_detection(
-    live_server, page: Page, mocked_authentification_user, visibilite_libelle: str
-):
-    """Test qu'un agent appartennant à la structure créatrice d'une fiche de détection peut voir la fiche quelque soit sa visibilité"""
-    fiche_detection = baker.make(
-        FicheDetection, visibilite=visibilite_libelle, createur=mocked_authentification_user.agent.structure
-    )
-    response = page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
-    assert response.status == 200
-    page.goto(f"{live_server.url}{reverse('fiche-liste')}")
-    expect(page.get_by_role("link", name=str(fiche_detection.createur))).to_be_visible()
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize("visibilite_libelle", [Visibilite.BROUILLON, Visibilite.LOCAL])
-def test_agent_not_in_structure_createur_cannot_view_fiche_detection_brouillon_or_local(
-    live_server, page: Page, mocked_authentification_user, visibilite_libelle: str
-):
-    """Test qu'un agent n'appartenant pas à la structure créatrice d'une fiche détection ne peut pas voir la fiche
-    si elle est en visibilité brouillon ou local"""
-    fiche_detection = baker.make(FicheDetection, visibilite=visibilite_libelle)
-    mocked_authentification_user.agent.structure = baker.make(Structure)
-    mocked_authentification_user.agent.save()
-    page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
-    expect(page.get_by_text("403 Forbidden")).to_be_visible()
-    page.goto(f"{live_server.url}{reverse('fiche-liste')}")
-    expect(page.get_by_role("link", name=str(fiche_detection.numero))).not_to_be_visible()
-
-
-@pytest.mark.django_db
-def test_agent_not_in_structure_createur_can_view_fiche_detection_nationale(
-    live_server, page: Page, fiche_detection: FicheDetection, mocked_authentification_user
-):
-    """Test qu'un agent n'appartenant pas à la structure créatrice d'une fiche détection peut voir la fiche
-    si elle est en visibilité national"""
-    _update_visibilite_fiche_detection(fiche_detection, Visibilite.NATIONAL)
-    page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
-    expect(page.get_by_role("heading", name=f"Fiche détection n° {str(fiche_detection.numero)}")).to_be_visible()
-    page.goto(f"{live_server.url}{reverse('fiche-liste')}")
-    expect(page.get_by_role("link", name=str(fiche_detection.numero))).to_be_visible()
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize("structure_ac", [MUS_STRUCTURE, BSV_STRUCTURE])
-def test_agent_ac_cannot_view_fiche_detection_brouillon(
-    live_server, page: Page, mocked_authentification_user, structure_ac: str
-):
-    """Test qu'un agent appartenant à l'AC ne peut pas voir une fiche détection en visibilité brouillon"""
-    fiche_detection = baker.make(FicheDetection, visibilite=Visibilite.BROUILLON)
-    mocked_authentification_user.agent.structure, _ = Structure.objects.get_or_create(
-        niveau1=AC_STRUCTURE, niveau2=structure_ac
-    )
-    mocked_authentification_user.agent.save()
-    page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
-    expect(page.get_by_text("403 Forbidden")).to_be_visible()
-    page.goto(f"{live_server.url}{reverse('fiche-liste')}")
-    expect(page.get_by_role("link", name=str(fiche_detection.numero))).not_to_be_visible()
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize("visibilite_libelle", [Visibilite.LOCAL, Visibilite.NATIONAL])
-@pytest.mark.parametrize("structure_ac", [MUS_STRUCTURE, BSV_STRUCTURE])
-def test_agent_ac_can_view_fiche_detection(
-    live_server,
-    page: Page,
-    fiche_detection: FicheDetection,
-    mocked_authentification_user,
-    visibilite_libelle: str,
-    structure_ac: str,
-):
-    """Test qu'un agent appartenant à l'AC peut voir une fiche détection en visibilité local ou national"""
-    _update_visibilite_fiche_detection(fiche_detection, visibilite_libelle)
-    mocked_authentification_user.agent.structure, _ = Structure.objects.get_or_create(
-        niveau1=AC_STRUCTURE, niveau2=structure_ac
-    )
-    mocked_authentification_user.agent.save()
-    page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
-    expect(page.get_by_role("heading", name=f"Fiche détection n° {str(fiche_detection.numero)}")).to_be_visible()
-    page.goto(f"{live_server.url}{reverse('fiche-liste')}")
-    expect(page.get_by_role("link", name=str(fiche_detection.numero))).to_be_visible()
+    page.wait_for_timeout(600)
+    fiche_detection = FicheDetection.objects.get()
+    assert fiche_detection.visibilite == Visibilite.LOCAL
 
 
 @pytest.mark.django_db
@@ -126,7 +39,8 @@ def test_cannot_update_fiche_detection_visibilite_by_other_structures(
 ):
     """Test que les agents n'appartenant pas à la structure créatrice d'une fiche de détection ne peuvent pas modifier la visibilité de cette fiche
     lorsqu'elle est en visibilité national."""
-    _update_visibilite_fiche_detection(fiche_detection, Visibilite.NATIONAL)
+    fiche_detection.visibilite = Visibilite.NATIONAL
+    fiche_detection.save()
     page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
     expect(page.get_by_text("Modifier la visibilité")).not_to_be_visible()
 
@@ -158,7 +72,8 @@ def test_agent_in_structure_createur_cannot_update_fiche_detection_visibilite(
 ):
     """Test qu'un agent appartenant à la structure créatrice d'une fiche détection
     ne peut pas modifier la visibilité de cette fiche si elle est en visibilité local ou national"""
-    _update_visibilite_fiche_detection(fiche_detection, visibilite_libelle)
+    fiche_detection.visibilite = visibilite_libelle
+    fiche_detection.save()
     fiche_detection.createur = mocked_authentification_user.agent.structure
     fiche_detection.save()
     page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
@@ -171,7 +86,8 @@ def test_agent_ac_can_update_fiche_detection_visibilite_local_to_national(
     live_server, page: Page, fiche_detection: FicheDetection, mocked_authentification_user, structure_ac: str
 ):
     """Test qu'un agent appartenant à l'AC peut modifier la visibilité d'une fiche détection de local à national"""
-    _update_visibilite_fiche_detection(fiche_detection, Visibilite.LOCAL)
+    fiche_detection.visibilite = Visibilite.LOCAL
+    fiche_detection.save()
     mocked_authentification_user.agent.structure, _ = Structure.objects.get_or_create(
         niveau1=AC_STRUCTURE, niveau2=structure_ac
     )
@@ -194,7 +110,8 @@ def test_agent_ac_can_update_fiche_detection_visibilite_national_to_local(
     live_server, page: Page, fiche_detection: FicheDetection, mocked_authentification_user, structure_ac: str
 ):
     """Test qu'un agent appartenant à l'AC peut modifier la visibilité d'une fiche détection de national à local"""
-    _update_visibilite_fiche_detection(fiche_detection, Visibilite.NATIONAL)
+    fiche_detection.visibilite = Visibilite.NATIONAL
+    fiche_detection.save()
     mocked_authentification_user.agent.structure, _ = Structure.objects.get_or_create(
         niveau1=AC_STRUCTURE, niveau2=structure_ac
     )
@@ -207,5 +124,17 @@ def test_agent_ac_can_update_fiche_detection_visibilite_national_to_local(
     page.get_by_role("button", name="Valider").click()
     expect(page.get_by_role("heading", name="La visibilité de la fiche détection a bien été modifiée")).to_be_visible()
     expect(page.get_by_text(Visibilite.LOCAL, exact=True)).to_be_visible()
+    fiche_detection.refresh_from_db()
+    assert fiche_detection.visibilite == Visibilite.LOCAL
+
+
+def test_publier_fiche_detection_from_btn(live_server, page: Page, mocked_authentification_user):
+    fiche_detection = FicheDetection.objects.create(
+        visibilite=Visibilite.BROUILLON, createur=mocked_authentification_user.agent.structure
+    )
+    page.goto(f"{live_server.url}{fiche_detection.get_absolute_url()}")
+    page.get_by_role("button", name="Publier").click()
+    expect(page.get_by_role("heading", name="La visibilité de la fiche détection a bien été modifiée")).to_be_visible()
+    expect(page.get_by_text(Visibilite.LOCAL)).to_be_visible()
     fiche_detection.refresh_from_db()
     assert fiche_detection.visibilite == Visibilite.LOCAL
